+++ src/main/java/com/thxforservice/counseling/controllers/CounselingAdminController.java
@@ -1,19 +1,27 @@
 package com.thxforservice.counseling.controllers;
 
+import com.thxforservice.counseling.constants.Status;
 import com.thxforservice.counseling.controllers.RequestGroupCounselingSave;
+import com.thxforservice.counseling.entities.Counseling;
 import com.thxforservice.counseling.exceptions.CounselingNotFoundException;
+import com.thxforservice.counseling.services.CounselingInfoService;
+import com.thxforservice.counseling.services.CounselingStatusService;
 import com.thxforservice.counseling.services.GroupCounselingDeleteService;
 import com.thxforservice.counseling.services.GroupCounselingSaveService;
 import com.thxforservice.counseling.validators.GroupCounselingValidator;
+import com.thxforservice.global.ListData;
 import com.thxforservice.global.Utils;
 import com.thxforservice.global.exceptions.BadRequestException;
 import com.thxforservice.global.rests.JSONData;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.springframework.validation.FieldError;
@@ -32,6 +40,8 @@ public class CounselingAdminController {
     private final HttpServletRequest request;
     private final GroupCounselingSaveService counselingSaveService;
     private final GroupCounselingValidator validator;
+    private final CounselingStatusService counselingStatusService;
+    private final CounselingInfoService counselingInfoService;
 
     private final Utils utils;
 
@@ -88,15 +98,36 @@ public class CounselingAdminController {
     // 집단 상담 E
 
     // 개별 상담 S
-    @Operation(summary = "개별 상담 신청 목록")
-    @GetMapping("/apply")
-    public JSONData privateApplyList() {
-        return null;
+    @Operation(summary = "상담사의 학생 예약 조회", method = "GET")
+    @GetMapping("/cs/list")
+    @PreAuthorize("hasAnyAuthority('COUNSELOR')")
+    public JSONData csList(CounselingSearch search) {
+
+        ListData<Counseling> data = counselingInfoService.getList(search);
+
+        return new JSONData(data);
     }
 
-    @Operation(summary = "개별 상담 신청 정보")
-    @GetMapping("/apply/{cSeq}")
-    public JSONData privateApplyInfo(@PathVariable("cSeq") Long cSeq) {
-        return null;
+    @Operation(summary = "상담사의 학생 예약 상세 조회", method = "GET")
+    @ApiResponse(responseCode = "201")
+    @GetMapping("/cs/info")
+    @PreAuthorize("hasAnyAuthority('COUNSELOR')")
+    public JSONData csInfo(@PathVariable("cSeq") Long cSeq) {
+
+        Counseling counseling = counselingInfoService.get(cSeq);
+
+        return new JSONData(counseling);
     }
+
+    @Operation(summary = "상담사의 학생 예약 상태 변경", method = "POST")
+    @PostMapping("/cs/status")
+    @PreAuthorize("hasAnyAuthority(('COUNSELOR'))")
+    public void CsChangeStatus(@Valid @RequestBody RequestCsChange form, Errors errors) {
+        if (errors.hasErrors()) {
+            throw new BadRequestException(utils.getErrorMessages(errors));
+        }
+        counselingStatusService.change(form.getCSeq(), Status.valueOf(form.getStatus()));
+    }
+    // 개별 상담 E
+
 }
