+++ src/main/java/com/thxforservice/counseling/controllers/CounselingController.java
@@ -2,9 +2,10 @@ package com.thxforservice.counseling.controllers;
 
 import com.thxforservice.counseling.constants.Status;
 import com.thxforservice.counseling.entities.Counseling;
-import com.thxforservice.counseling.entities.GroupProgram;
-import com.thxforservice.counseling.repositories.CounselingRepository;
-import com.thxforservice.counseling.services.*;
+import com.thxforservice.counseling.services.CounselingApplyService;
+import com.thxforservice.counseling.services.CounselingCancelService;
+import com.thxforservice.counseling.services.CounselingInfoService;
+import com.thxforservice.counseling.services.CounselingStatusService;
 import com.thxforservice.counseling.validators.CounselingValidator;
 import com.thxforservice.global.ListData;
 import com.thxforservice.global.Utils;
@@ -13,12 +14,9 @@ import com.thxforservice.global.rests.JSONData;
 import com.thxforservice.member.MemberUtil;
 import com.thxforservice.member.entities.Member;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import jakarta.ws.rs.Path;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -38,6 +36,7 @@ public class CounselingController {
      * 1. 개인 상담 신청  - POST /apply
      * - 개인 상담 목록 조회 ( 다중 )
      * - 개인 상담 목록 조회 ( 단일 )
+     * - 개인 상담 예약 취소
      * -------상담사 -------
      * 1. 개인 상담 목록 조회 ( 다중 )
      * 2. 개인 상담 목록 조회 ( 단일 )
@@ -56,6 +55,7 @@ public class CounselingController {
     private final CounselingInfoService infoService;
     private final MemberUtil memberUtil;
     private final CounselingStatusService statusService;
+    private final CounselingCancelService cancelService;
 
     @Operation(summary = "개인 상담 신청", method = "POST")
     @ApiResponse(responseCode = "201")
@@ -106,41 +106,10 @@ public class CounselingController {
     @Operation(summary = "개인 상담 예약 취소", method = "POST")
     @ApiResponse(responseCode = "201")
     @PostMapping("/cancel/{cSeq}")
-    public JSONData cancel (@PathVariable("cSeq") Long cSeq) {
-       Counseling item = infoService.cancel(cSeq);
+    public JSONData cancel(@PathVariable("cSeq") Long cSeq) {
+        Counseling item = cancelService.cancel(cSeq);
 
         return new JSONData(item);
-    }
-
-    @Operation(summary = "상담사의 학생 예약 조회", method = "GET")
-    @GetMapping("/cs/list")
-    @PreAuthorize("hasAnyAuthority('COUNSELOR')")
-    public JSONData csList(CounselingSearch search) {
-
-        ListData<Counseling> data = infoService.getList(search);
-
-        return new JSONData(data);
-    }
-
-    @Operation(summary = "상담사의 학생 예약 상세 조회", method = "GET")
-    @ApiResponse(responseCode = "201")
-    @GetMapping("/cs/info")
-    @PreAuthorize("hasAnyAuthority('COUNSELOR')")
-    public JSONData csInfo(@PathVariable("cSeq") Long cSeq) {
-
-        Counseling counseling = infoService.get(cSeq);
-
-        return new JSONData(counseling);
-    }
-
-    @Operation(summary = "상담사의 학생 예약 상태 변경", method = "POST")
-    @PostMapping("/cs/status")
-    @PreAuthorize("hasAnyAuthority(('COUNSELOR'))")
-    public void CsChangeStatus(@Valid @RequestBody RequestCsChange form, Errors errors){
-        if(errors.hasErrors()) {
-            throw new BadRequestException(utils.getErrorMessages(errors));
-        }
-        statusService.change(form.getCSeq(), Status.valueOf(form.getStatus()));
     }
 
     @Operation(summary = "상담사 평점 - 개인 상담, 집단 상담")
