+++ src/main/java/com/thxforservice/counseling/entities/GroupProgram.java
@@ -1,19 +1,17 @@
 package com.thxforservice.counseling.entities;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.thxforservice.global.entities.BaseMemberEntity;
 import com.thxforservice.counseling.constants.ProgramStatus;
+import com.thxforservice.global.entities.BaseMemberEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Min;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.ColumnDefault;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @Entity
